requirements.txt
python-dotenv
kucoin-python
ta
pandas
python-telegram-bot
